move ErrorStatusCode to its own module
