+++ dropshot/src/lib.rs
@@ -852,6 +852,7 @@ mod api_description;
 mod body;
 mod config;
 mod error;
+mod error_status_code;
 mod extractor;
 mod from_map;
 mod handler;
@@ -893,14 +894,14 @@ pub use config::ConfigTls;
 pub use config::HandlerTaskMode;
 pub use config::RawTlsConfig;
 pub use dtrace::ProbeRegistration;
-pub use error::ClientErrorStatusCode;
-pub use error::ErrorStatusCode;
 pub use error::HttpError;
 pub use error::HttpErrorResponseBody;
-pub use error::InvalidClientErrorStatusCode;
-pub use error::InvalidErrorStatusCode;
-pub use error::NotAClientError;
-pub use error::NotAnError;
+pub use error_status_code::ClientErrorStatusCode;
+pub use error_status_code::ErrorStatusCode;
+pub use error_status_code::InvalidClientErrorStatusCode;
+pub use error_status_code::InvalidErrorStatusCode;
+pub use error_status_code::NotAClientError;
+pub use error_status_code::NotAnError;
 pub use extractor::ExclusiveExtractor;
 pub use extractor::ExtractorMetadata;
 pub use extractor::MultipartBody;
